+++ packages/plugin-farcaster/src/common/constants.ts
@@ -1,0 +1,6 @@
+export const FARCASTER_SERVICE_NAME = 'farcaster';
+export const FARCASTER_SOURCE = 'farcaster';
+export const DEFAULT_MAX_CAST_LENGTH = 320;
+export const DEFAULT_POLL_INTERVAL = 120; // 2 minutes
+export const DEFAULT_POST_INTERVAL_MIN = 90; // 1.5 hours
+export const DEFAULT_POST_INTERVAL_MAX = 180; // 3 hours
