+++ packages/plugin-farcaster/src/common/prompts.ts
