+++ packages/plugin-farcaster/src/common/utils.ts
