--- packages/plugin-farcaster/src/constants.ts
@@ -1,2 +1,0 @@
-export const FARCASTER_SERVICE_NAME = 'farcaster';
-export const FARCASTER_SOURCE = 'farcaster';
