+++ packages/plugin-farcaster/src/service.ts
@@ -1,66 +1,11 @@
 import { elizaLogger, logger, Service, UUID, type IAgentRuntime } from '@elizaos/core';
-import { Configuration, NeynarAPIClient } from '@neynar/nodejs-sdk';
-import { FarcasterClient } from './client';
-import { FARCASTER_SERVICE_NAME } from './constants';
-import { validateFarcasterConfig, type FarcasterConfig } from './environment';
-import { FarcasterInteractionManager } from './interactions';
-import { FarcasterPostManager } from './post';
+import { FARCASTER_SERVICE_NAME } from './common/constants';
+import { FarcasterAgentManager } from './managers/agent';
+import { hasFarcasterEnabled, validateFarcasterConfig } from './common/environment';
 
-/**
- * A manager that orchestrates all Farcaster operations:
- * - client: base operations (Neynar client, hub connection, etc.)
- * - posts: autonomous posting logic
- * - interactions: handling mentions, replies, likes, etc.
- */
-class FarcasterManager {
-  client: FarcasterClient;
-  posts: FarcasterPostManager;
-  interactions: FarcasterInteractionManager;
-  private signerUuid: string;
-
-  constructor(runtime: IAgentRuntime, farcasterConfig: FarcasterConfig) {
-    const cache = new Map<string, any>();
-    this.signerUuid = runtime.getSetting('FARCASTER_NEYNAR_SIGNER_UUID')!;
-
-    const neynarConfig = new Configuration({
-      apiKey: runtime.getSetting('FARCASTER_NEYNAR_API_KEY')!,
-    });
-
-    const neynarClient = new NeynarAPIClient(neynarConfig);
-
-    this.client = new FarcasterClient({
-      runtime,
-      ssl: true,
-      url: runtime.getSetting('FARCASTER_HUB_URL') ?? 'hub.pinata.cloud',
-      neynar: neynarClient,
-      signerUuid: this.signerUuid,
-      cache,
-      farcasterConfig,
-    });
-
-    elizaLogger.success('Farcaster Neynar client initialized.');
-
-    this.posts = new FarcasterPostManager(this.client, runtime, this.signerUuid, cache);
-
-    this.interactions = new FarcasterInteractionManager(
-      this.client,
-      runtime,
-      this.signerUuid,
-      cache
-    );
-  }
-
-  async start() {
-    await Promise.all([this.posts.start(), this.interactions.start()]);
-  }
-
-  async stop() {
-    await Promise.all([this.posts.stop(), this.interactions.stop()]);
-  }
-}
 export class FarcasterService extends Service {
-  private static instance: FarcasterService | undefined;
-  private managers = new Map<UUID, FarcasterManager>();
+  private static instance?: FarcasterService;
+  private managers = new Map<UUID, FarcasterAgentManager>();
   static serviceType: string = FARCASTER_SERVICE_NAME;
   readonly capabilityDescription = 'The agent is able to send and receive messages on farcaster';
 
@@ -71,6 +16,44 @@ export class FarcasterService extends Service {
     return FarcasterService.instance;
   }
 
+  // Called to start a single Farcaster service
+  static async start(runtime: IAgentRuntime): Promise<Service> {
+    const service = FarcasterService.getInstance();
+    let manager = service.managers.get(runtime.agentId);
+
+    if (manager) {
+      logger.warn('Farcaster service already started', runtime.agentId);
+      return service;
+    }
+
+    if (!hasFarcasterEnabled(runtime)) {
+      logger.debug('Farcaster service not enabled', runtime.agentId);
+      return service;
+    }
+
+    const farcasterConfig = validateFarcasterConfig(runtime);
+    manager = new FarcasterAgentManager(runtime, farcasterConfig);
+    service.managers.set(runtime.agentId, manager);
+    await manager.start();
+
+    elizaLogger.log('Farcaster client started', runtime.agentId);
+    return service;
+  }
+
+  // Called to stop a single Farcaster service
+  static async stop(runtime: IAgentRuntime): Promise<void> {
+    const service = FarcasterService.getInstance();
+    let manager = service.managers.get(runtime.agentId);
+    if (manager) {
+      await manager.stop();
+      service.managers.delete(runtime.agentId);
+      elizaLogger.log('Farcaster client stopped', runtime.agentId);
+    } else {
+      logger.warn('Farcaster service not running', runtime.agentId);
+    }
+  }
+
+  // Called to stop all Farcaster services
   async stop(): Promise<void> {
     logger.log('Stopping ALL Farcaster services');
     for (const manager of Array.from(this.managers.values())) {
@@ -80,36 +63,6 @@ export class FarcasterService extends Service {
       } catch (error) {
         logger.error('Error stopping Farcaster service', agentId, error);
       }
-    }
-  }
-
-  static async start(runtime: IAgentRuntime): Promise<Service> {
-    const service = FarcasterService.getInstance();
-    let manager = service.managers.get(runtime.agentId);
-
-    if (manager) {
-      logger.warn('Farcaster service already started');
-      return service;
-    }
-
-    const farcasterConfig = validateFarcasterConfig(runtime);
-    manager = new FarcasterManager(runtime, farcasterConfig);
-    service.managers.set(runtime.agentId, manager);
-    await manager.start();
-
-    elizaLogger.log('Farcaster client started');
-    return service;
-  }
-
-  /** Stop service connection */
-  static async stop(runtime: IAgentRuntime): Promise<void> {
-    const service = FarcasterService.getInstance();
-    let manager = service.managers.get(runtime.agentId);
-    if (manager) {
-      await manager.stop();
-      service.managers.delete(runtime.agentId);
-    } else {
-      logger.warn('Farcaster service not started', runtime.agentId);
     }
   }
 }
